phase-2/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project]
name = "todo-backend"
version = "0.1.0"
description = "Hackathon II Todo Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
    "email-validator>=2.0.0",
]



[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]

[tool.uv.test]
command = "pytest"
args = ["-q"]
